% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carstats.R
\docType{data}
\name{carstats}
\alias{carstats}
\title{Data from carstats}
\format{A \code{\link{data.table}} object with 406 rows and 9 columns.}
\usage{
carstats
}
\description{
Copy from the carstats sample dataset available in Cornerstone.
}
\keyword{datasets}
