% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rundata}
\alias{rundata}
\title{Voltage of 26 Different Runs}
\format{
A \code{\link{data.table}} object with 7826 observations and 3 variables.
  The discrete variable 'FileName' has 26 levels corresponding to the file name of the source data
  from Cornerstone. The remaining other two variables are continuous.
}
\source{
Cornerstone sample dataset
}
\description{
The datasets contains 26 different runs in each case measuring the voltage over time. The S-shaped
  curves are different for each run.
}
