% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showVersions.R
\name{showVersions}
\alias{showVersions}
\title{Show Versions of R and CornerstoneR}
\usage{
showVersions(return.results = FALSE)
}
\arguments{
\item{return.results}{[\code{logical(1)}]\cr
If \code{FALSE} the function returns \code{TRUE} invisibly.
If \code{TRUE}, it returns a \code{\link{list}} of results.
Default is \code{FALSE}.}
}
\value{
Logical [\code{TRUE}] invisibly and outputs to Cornerstone or, 
  if \code{return.results = TRUE}, \code{\link{list}} of 
  resulting \code{\link{data.frame}} objects:
  \item{versions}{
    Dataset with versions of R and CornerstoneR.
  }
}
\description{
Write the versions of R and CornerstoneR in a Cornerstone dataset.
}
\examples{
  res = showVersions(return.results = TRUE)
  res$versions
}
