% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{titanic}
\alias{titanic}
\title{Survival of Passengers on the Titanic}
\format{
A \code{\link{data.table}} object with 2201 observations and 4 variables.
  The discrete variables and their levels are as follows:
  \itemize{
    \item{Class: }{1st, 2nd, 3rd, Crew}
    \item{Age: }{Child, Adult}
    \item{Sex: }{Male, Female}
    \item{Survived: }{Yes, No}
  }
}
\source{
Cornerstone sample dataset
}
\description{
This data set provides information on the fate of passengers on the fatal maiden voyage 
  of the ocean liner Titanic, summarized according to economic status (Class), Age, Sex, and Survival.
}
\references{
Dawson, Robert J. MacG. (1995), The Unusual Episode Data Revisited. 
  \emph{Journal of Statistics Education}, \strong{3}. 
  \url{ http://jse.amstat.org/v3n3/datasets.dawson.html}
}
