% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carstats}
\alias{carstats}
\title{Data from carstats}
\format{
A \code{\link{data.table}} object with 406 observations and 9 variables.
  The variables and their scale of measurement are as follows:
  \itemize{
    \item{Model: }{nominal}
    \item{Origin: }{nominal}
    \item{MPG: }{interval}
    \item{Cylinders: }{ordinal}
    \item{Displacement: }{interval}
    \item{Horsepower: }{interval}
    \item{Weight: }{interval}
    \item{Acceleration: }{interval}
    \item{Model.Year: }{interval}
  }
}
\source{
Cornerstone sample dataset
}
\description{
Dataset of different cars and various values.
}
