% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mosaicPlot.R
\name{mosaicPlot}
\alias{mosaicPlot}
\title{Mosaic Plot}
\usage{
mosaicPlot(dataset = cs.in.dataset(), preds = cs.in.predictors(),
  resps = cs.in.responses())
}
\arguments{
\item{dataset}{[\code{\link{data.frame}}]\cr
Dataset with named columns. The names correspond to predictors and responses.}

\item{preds}{[\code{character}]\cr
Character vector of predictors.}

\item{resps}{[\code{character}]\cr
Character vector of responses.}
}
\description{
Plot extended mosaic via \code{\link[vcd]{mosaic}}.
}
\examples{
# Draw mosaic plot from 'Titanic' data:
mosaicPlot(as.data.frame(Titanic), c("Class", "Sex", "Age", "Survived"), "Freq")
}
